Fix displaying of models (Dragon works!)

+++ DirectX12FromScratch/include/MiniEngine/D3D12/D3D12RenderableModel.h
@@ -33,6 +33,6 @@ namespace MiniEngine
 		D3D12IndexBuffer    *_indexBuffer;
 
 		D3D12Vertex         *_vertexData;
-		unsigned short      *_indexData;
+		unsigned int        *_indexData;
 	};
 }
